requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
python-dotenv==1.0.1